Fixes #841 

**Summary :-**  

Replaced naive Disjoint-set data structure and search algorithm with a more optimal one that noticeably improves performance with a slight decrease in memory overhead.

**Benchmark Results:-**
|E| = 5 * |V|

For |V| = 1000, 

Old Implementation:              2.777 ms (17204 allocations: 790.56 KiB)
Optimized Implementation:   1.529 ms (17981 allocations: 771.64 KiB)

For |V| = 10000, 

Old Implementation:            164.897 ms (281392 allocations: 9.40 MiB)
Optimized Implementation:   28.465 ms (278255 allocations: 9.05 MiB)

For |V| = 100000, 

Old Implementation:                13.637 s (2957230 allocations: 92.42 MiB)
Optimized Implementation:   666.549 ms (3044722 allocations: 90.71 MiB)