@SohamTamba 2018-02-16
I also made an orphan [branch](https://github.com/SohamTamba/LightGraphs.jl/tree/kruskal_test) that contains the current implementation exposed as kruskal_mst_old and the proposed one as kruskal_mst. 

This branch also contains a benchmark [script](url) that one can run to try and reproduce the trend of results I have got. I'd appreciate if someone can verify my results on a different machine.