@SohamTamba 2018-02-17
I've made the suggested changes and squashed to one commit.